requirements.txt
azure-storage-blob
azure-storage-file-share
azure-storage-queue
azure-data-tables
